+++ src/front/component/buttomInformation/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";  
+import 'bootstrap/dist/css/bootstrap.min.css';
+
+
+function ButtomInformation() {
+    return (
+
+        <button type="button" class="btn btn-secondary">Consulta con nosotros</button>
+    )
+}
+
+export default ButtomInformation;
